requirements.txt
fastapi
sqlalchemy
pydantic
pydantic[email]
pyjwt
cryptography
argon2-cffi
python-multipart
uvicorn
gunicorn ; platform_system!="Windows"
asyncpg
redis
icecream
httpx
python-dotenv
aiosmtplib
uvloop ; platform_system!="Windows"
jinja2
pytz
orjson
pydantic-settings